Import clusters results generated by SPeDE

Reformat the table output from the analysis of raw Bruker MALDI Biotyper spectra by the SPeDE tool from Dumolin et al. (2019) to be consistent with the Strejcek et al. (2018) procedure followed in the maldipickr package.

Usage

import_spede_clusters(path)

Arguments

path

Path to the comma separated table generated by SPeDE

Value

A tibble with the following columns:

name: a character denoting the spectra name (all spaces, dashes and dots are replaced by underscores "_" in SPeDE)

membership: integers stating the cluster number to which the spectra belong to. It starts from 1 to c, the total number of clusters.

cluster_size: integers indicating the total number of spectra in the corresponding cluster.

quality: a character indicating the spectra quality category by SPeDE, out of GREEN, ORANGE and RED.

is_reference: a logical indicating whether the corresponding spectra is a reference spectra of the cluster.

References

Dumolin C, Aerts M, Verheyde B, Schellaert S, Vandamme T, Van Der Jeugt F, De Canck E, Cnockaert M, Wieme AD, Cleenwerck I, Peiren J, Dawyndt P, Vandamme P, & Carlier A. (2019). "Introducing SPeDE: High-Throughput Dereplication and Accurate Determination of Microbial Diversity from Matrix-Assisted Laser Desorption–Ionization Time of Flight Mass Spectrometry Data". MSystems 4(5). doi:10.1128/msystems.00437-19.

See also

Examples

# Reformat the output from SPeDE table
# https://github.com/LM-UGent/SPeDE
import_spede_clusters(
  system.file("spede.csv", package = "maldipickr")
)
#> # A tibble: 6 × 5
#>   name         membership cluster_size quality is_reference
#>   <chr>             <dbl>        <int> <chr>   <lgl>       
#> 1 species1_G2           1            1 GREEN   TRUE        
#> 2 species2_E11          2            2 ORANGE  FALSE       
#> 3 species2_E12          2            2 GREEN   TRUE        
#> 4 species3_F7           3            1 GREEN   TRUE        
#> 5 species3_F8           4            2 ORANGE  FALSE       
#> 6 species3_F9           4            2 GREEN   TRUE