maldipickr: Dereplicate and Cherry-Pick Mass Spectrometry Spectra

Convenient wrapper functions for the analysis of matrix-assisted laser desorption/ionization-time-of-flight (MALDI-TOF) spectra data in order to select only representative spectra (also called cherry-pick). The package covers the preprocessing and dereplication steps (based on Strejcek, Smrhova, Junkova and Uhlik 2018 doi:10.3389/fmicb.2018.01294 ) needed to cluster MALDI-TOF spectra before the final cherry-picking step. It enables the easy exclusion of spectra and/or clusters to accommodate complex cherry-picking strategies. Alternatively, cherry-picking using taxonomic identification MALDI-TOF data is made easy with functions to import inconsistently formatted reports.

See also

Author

Maintainer: Charlie Pauvert cpauvert@ukaachen.de (ORCID) [copyright holder]

Other contributors:

David Wylensek (ORCID) [contributor]

Selina Nüchtern [contributor]

Thomas Clavel (ORCID) [contributor, funder, copyright holder]