Cherry-pick Bruker MALDI Biotyper spectra

Using the clusters information, and potential additional metadata as external criteria, spectra are labelled as to be picked for each cluster. Note that some spectra and therefore clusters can be explicitly removed from the picking decision using regular expressions, if they have been previously picked. If no metadata are provided, the reference spectra of each cluster will be picked.

Usage

pick_spectra(
  cluster_df,
  metadata_df = NULL,
  criteria_column = NULL,
  is_descending_order = TRUE,
  discard_regex = NULL,
  only_show_discarded = FALSE,
  is_sorted = FALSE
)

Arguments

cluster_df

A tibble with clusters information from the similarity_to_clusters or the import_spede_clusters function.

metadata_df

Optional tibble with relevant metadata to guide the picking process (e.g., OD600).

criteria_column

Optional character indicating the column in metadata_df to be used as a criteria.

is_descending_order

Optional logical indicating whether to sort in descending order or not.

discard_regex

Optional regex to remove spectra and therefore clusters from the picking decision (e.g., if they have been picked already).

only_show_discarded

Optional logical to be used in conjunction to discard_regex to only print the names of the spectra to be discarded as a verification step.

is_sorted

Optional logical to indicate that the cluster_df is already sorted by cluster based on (usually multiple) internal criteria to pick the first of each cluster. This flag is overriden if a metadata_df is provided.

Value

A tibble with as many rows as cluster_df with an additional logical column named to_pick to indicate whether the colony associated to the spectra should be picked. If metadata_df is provided, then additional columns from this tibble are added to the returned tibble. When the discard_regex is used, the to_pick column of the tibble for the discarded spectra displays FALSE. When only_show_discarded is used, it returns $0$ and print out the list of spectra to be removed.

Examples

# Load a toy example of a tibble of clusters created by the `similarity_to_clusters`
#  function.
clusters <- readRDS(system.file("clusters_tibble.RDS", package = "maldipickr"))

# By default and if no other metadata are provided, the function
#  picks the reference spectrum of each clusters.
pick_spectra(clusters)
#> # A tibble: 6 × 7
#>   name         membership cluster_size   SNR peaks is_reference to_pick
#>   <chr>             <int>        <int> <dbl> <dbl> <lgl>        <lgl>  
#> 1 species1_G2           1            4  5.09    21 FALSE        FALSE  
#> 2 species2_E11          2            2  5.54    22 FALSE        FALSE  
#> 3 species2_E12          2            2  5.63    23 TRUE         TRUE   
#> 4 species3_F7           1            4  4.89    26 FALSE        FALSE  
#> 5 species3_F8           1            4  5.56    25 TRUE         TRUE   
#> 6 species3_F9           1            4  5.40    25 FALSE        FALSE  

# Simulate OD600 values with uniform distribution
#  for each of the colonies we measured with
#  the Bruker MALDI Biotyper
set.seed(104)
metadata <- dplyr::transmute(clusters, name = name, OD600 = runif(n = nrow(clusters)))
metadata
#> # A tibble: 6 × 2
#>   name         OD600
#>   <chr>        <dbl>
#> 1 species1_G2  0.364
#> 2 species2_E11 0.772
#> 3 species2_E12 0.735
#> 4 species3_F7  0.973
#> 5 species3_F8  0.740
#> 6 species3_F9  0.201

# Pick the spectra based on the highest OD600 value per cluster
pick_spectra(clusters, metadata, "OD600")
#> # A tibble: 6 × 8
#>   name         membership cluster_size   SNR peaks is_reference OD600 to_pick
#>   <chr>             <int>        <int> <dbl> <dbl> <lgl>        <dbl> <lgl>  
#> 1 species1_G2           1            4  5.09    21 FALSE        0.364 FALSE  
#> 2 species2_E11          2            2  5.54    22 FALSE        0.772 TRUE   
#> 3 species2_E12          2            2  5.63    23 TRUE         0.735 FALSE  
#> 4 species3_F7           1            4  4.89    26 FALSE        0.973 TRUE   
#> 5 species3_F8           1            4  5.56    25 TRUE         0.740 FALSE  
#> 6 species3_F9           1            4  5.40    25 FALSE        0.201 FALSE  

# Say that spectra 'species2_E11' was picked before in a previous experiment
#  we do not want to pick a cluster where species2_E11 is included to limit redundancy
# The function can peform a dry run to indicate which spectra would be removed
pick_spectra(clusters, discard_regex = "E11", only_show_discarded = TRUE)
#> Warning: Returning more (or less) than 1 row per `summarise()` group was deprecated in
#> dplyr 1.1.0.
#> ℹ Please use `reframe()` instead.
#> ℹ When switching from `summarise()` to `reframe()`, remember that `reframe()`
#>   always returns an ungrouped data frame and adjust accordingly.
#> ℹ The deprecated feature was likely used in the maldipickr package.
#>   Please report the issue at <https://github.com/ClavelLab/maldipickr/issues>.
#> The following spectra are discarded:
#> species2_E11 species2_E12 
#> [1] 0
# Once checked, the real dereplication can be performed.
pick_spectra(clusters, discard_regex = "E11")
#> # A tibble: 6 × 7
#>   name         membership cluster_size   SNR peaks is_reference to_pick
#>   <chr>             <int>        <int> <dbl> <dbl> <lgl>        <lgl>  
#> 1 species1_G2           1            4  5.09    21 FALSE        FALSE  
#> 2 species2_E11         NA           NA NA       NA NA           FALSE  
#> 3 species2_E12         NA           NA NA       NA NA           FALSE  
#> 4 species3_F7           1            4  4.89    26 FALSE        FALSE  
#> 5 species3_F8           1            4  5.56    25 TRUE         TRUE   
#> 6 species3_F9           1            4  5.40    25 FALSE        FALSE